requirements.txt
requests
firebase_admin
joblib
python-dotenv